cookie fini, confirmation de la suppression d'un element ok, suppression d'une liste(toDoList) ok

+++ app/View/ToDos/tasks.ctp
@@ -61,7 +61,7 @@
 					</div>
 				</div>
 
-	            <a ng-href="{{base}}/Tasks/delete_task/<?php echo $idToDo ?>/{{value.Task.id}}" class="btn-delete">
+                                <a ng-href="{{base}}/Tasks/delete_task/<?php echo $idToDo ?>/{{value.Task.id}}" class="btn-delete" onclick="return confirm('Voulez-Vous supprimer l\'élément {{value.Task.id}} ?\n Cette action est irreversible.')">
 					<span class="glyphicon glyphicon-minus"></span>
 				</a>
             </label>
